+++ aura/__init__.py
@@ -1,3 +1,3 @@
 """Aura language toolchain."""
 
-__version__ = "0.1.0a15"
+__version__ = "0.1.0a16"